+++ scorched/search.py
@@ -378,7 +378,8 @@ class BaseSearch(object):
     option_modules = ('query_obj', 'filter_obj', 'paginator',
                       'more_like_this', 'highlighter', 'postings_highlighter',
                       'faceter', 'grouper', 'sorter', 'facet_querier',
-                      'field_limiter', 'parser', 'pivoter')
+                      'debugger', 'requesthandler', 'field_limiter', 'parser',
+                      'pivoter')
 
     def _init_common_modules(self):
         self.query_obj = LuceneQuery(u'q')
@@ -391,6 +392,8 @@ class BaseSearch(object):
         self.pivoter = FacetPivotOptions()
         self.grouper = GroupOptions()
         self.sorter = SortOptions()
+        self.debugger = DebugOptions()
+        self.requesthandler = RequestHandlerOption()
         self.field_limiter = FieldLimitOptions()
         self.facet_querier = FacetQueryOptions()
 
@@ -492,6 +495,16 @@ class BaseSearch(object):
     def paginate(self, start=None, rows=None):
         newself = self.clone()
         newself.paginator.update(start, rows)
+        return newself
+
+    def debug(self):
+        newself = self.clone()
+        newself.debugger.update(True)
+        return newself
+
+    def set_requesthandler(self, handler):
+        newself = self.clone()
+        newself.requesthandler.update(handler)
         return newself
 
     def sort_by(self, field):
@@ -1041,6 +1054,48 @@ class SortOptions(Options):
                 "%s %s" % (field, order) for order, field in self.fields)}
         else:
             return {}
+
+
+class DebugOptions(Options):
+    # XXX should be changed to 'debug' added in 4.0
+    # https://wiki.apache.org/solr/CommonQueryParameters#Debugging
+    option_name = "debugQuery"
+
+    def __init__(self, original=None):
+        if original is None:
+            self.debug = False
+        else:
+            self.debug = original.debug
+
+    def update(self, debug):
+        self.debug = debug
+
+    def options(self):
+        if self.debug:
+            return {"debugQuery": True}
+        else:
+            return {}
+
+
+class RequestHandlerOption(Options):
+    option_name = "qt"
+
+    def __init__(self, original=None):
+        if original is None:
+            # XXX 'standard' is deprecated
+            # https://wiki.apache.org/solr/SolrRequestHandler#Old_handleSelect.3Dtrue_Resolution_.28qt_param.29
+            self.handler = None
+        else:
+            self.handler = original.handler
+
+    def update(self, handler):
+        self.handler = handler
+
+    def options(self):
+        ret = {}
+        if self.handler:
+            ret = {"qt": self.handler}
+        return ret
 
 
 class FieldLimitOptions(Options):